refactor(core): rework config handling
# Goal

The goal is to improve the maintainability and testability of the config handling part. This is necessary because of the nightmare represented by the current setup when there is the need to put our hands in it.

In order to achieve the goal, multiple objectives have been defined:
* Decouple the config properties definition from the infrastructure-related stuffs (i.e. config file serialization)
* Get rid of the [Jakarta Beans Validator framework](https://beanvalidation.org/)
* Prepare the removal of the [Configurate framework](https://github.com/SpongePowered/Configurate)

## Why removing Jakarta Beans Validator

This solution seems amazing at first glance. In fact, the idea of validating beans with annotations sounds wonderful because it is a simple and intuitive way of enforcing constraints in applications.

However, disillusionment starts to hit as long as you begin to be strict about your needs and want the framework to fit them. You may even face the need to rely on a complementary library like [conditional-validator](https://github.com/microtweak/conditional-validator) in order to support some edge cases like "valid when" logics.

A similar disappointment hit when starting to rely more and more on annotations. You realize that only few basic annotations are provided by default, thus requiring you to define custom ones for advanced cases. On top of that, you will have to implement the `ConstraintValidator` interface and then associate it to the corresponding annotation you just created. Maybe overkill when performing an equivalent check with the [Apache Commons Lang library](https://github.com/apache/commons-lang) may only necessitate one instruction and is therefore far simpler.

But finally what convinced me to definitively stop using the framework is when I realized what was lost because of its major limitation: the fact validation can only be performed post-instantiation, thus losing the strict guarantee to have valid instances in the system. Stated another way: we are unable to bring all benefits of the [value object pattern](https://martinfowler.com/bliki/ValueObject.html). As a consequence, we will always have the fear to introduce unvalidated instances in the system without the possibility to completely mitigate this risk. That may sounds for you like nothing really serious, but you would be wrong.

Imagine an `Email` class referenced as parameter of several exposed methods of your library. Imagine as well that instantiation responsibility of this class is delegated to the library's consumer. Because we rely on the Jakarta Beans Validator framework for the validation, the implication is that the consumer will have to call somehow a validator. But let's say the consumer is able to find a way to validate the newly created `Email` instance... What if it just forgot to validate the bean before calling one of the library's methods? Can we really trust users? At this stage of thinking, you start to see that there is a risk of facing unexpected issues and even triggering bugs with potentially harmful or even catastrophic consequences. So you want to add tests to the multiple methods relying on this type for their parameters... And you end-up executing same tests you already wrote for the `Email` class itself at multiple places (that's without telling that you risk as well to forget some tests on some methods...).

Fair enough, but we can avoid such testing mess by hiding the `Email`'s constructor(s) and instead relying on the Factory pattern. The `EmailFactory` will then be responsible of instantiating the `Email` class and, just after that, calling the validator. Amazing! We no longer have to test everything everywhere. Additionally, the user experience is not bad since the validation is abstracted. But we now have yet another thing becoming more complex because of the use of the Jakarta Beans Validator framework... Whereas we can just put checks at constructor level... And we have the strongest possible type that we want, thus completely mitigating the initial risk of having unvalidated instances in the system.

I can continue by telling that there is a way stronger coupling with the framework on tests side than in production code (you need a validator in order to validate the annotations themselves), thus making tests creation, maintenance and evolution harder... But it will just be seen as the final blow against the framework usage.

The framework is paradoxal: it gives the illustion of a clean and simple solution, but in the end it may force you to make things so complex for nothing.

Maybe the framework can find its place in some specific cases and/or in simple applications (especially when knowing that frameworks like Spring Boot supports constraint annotations out-of-the-box)... But even in these cases I will not consider it since relying on Apache Commons Lang library or even implementing your own adhoc validation logic is better. It's trivial, efficient, flexible and allows you to define strong types, thus reducing risks of facing bugs because of unvalidated instances flowing into your system.

## Why removing Configurate

The major motivation for using such solution was for the round-robin serialization support for config files. Basically, it's the ability to deserialize a config file and then reserialize it back with the expection to get the exact same content in the end. That's something interesting when additing in the course of the process some data transformations so that it's possible to automatically migrate config files in existing deployments. Doing so avoid server administrators to have to do anything else than downloading and installing the new plugin's version.

However, such thing can be performed even with just the snakeyaml library, thus avoiding us the need to know and manipulate a dependency unnecessarily. Furthermore, I think the solution lack plus-value evidences over well-known, comprehensive and long time battle-tested solutions such as snakeyaml and Jackson tools.

Another motivation is to get rid of the HOCON format in order to come back to the plain old YAML format, convenient to use and well known in the Bukkit world even if imperfect. As a recall, the HOCON format was the one used because of limitations with the YAML implementation. However, the big drawback of the HOCON format is that properties ended up being shuffled at serialization time... Clearly not something amazing for the administrator nor what I wanted to propose them.

In short: it brings complexity, has limitations and bugs, is specific to Bukkit/Sponge systems and finally force to rely on the HOCON format with an imperfect implementation.